fix: dynamic route 삭제
